pyproject.toml
[project]
name = "pdf-summarizer"
version = "0.1.0"
description = ""
authors = [
    {name = "JooHeonJeong",email = "cherish8513@gihub.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "transformers (>=4.52.1,<5.0.0)",
    "hf-xet (>=1.1.2,<2.0.0)",
    "accelerate (>=1.7.0,<2.0.0)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "pdf_summarizer", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
